Group serialized client state structs.
