+++ Assets/Scripts/ClientState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b8daf981a9a5f7dbb82eb3a955ac10ed
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
