+++ Assets/Scripts/XRInputHelperComponent.cs
@@ -21,14 +21,6 @@ public class BoolArrayHelper
     }
 }
 
-[Serializable]
-public class ButtonInputData
-{
-    public List<int> buttonHeld = new List<int>();
-    public List<int> buttonUp = new List<int>();
-    public List<int> buttonDown = new List<int>();
-}
-
 public class XRInputHelper
 {
     XRIDefaultInputActions _inputActions;
@@ -109,30 +101,6 @@ public class XRInputHelper
     private void RightSelectCallback(InputAction.CallbackContext obj)
     {
         ButtonPressReleaseCallback(3, obj.performed);
-    }
-}
-
-[System.Serializable]
-public class AvatarData
-{
-    public PoseData root = new PoseData();
-    public PoseData[] hands = new PoseData[]
-    {
-        new PoseData(),
-        new PoseData()
-    };
-}
-
-[Serializable]
-public class PoseData
-{
-    public List<float> position = new List<float>(3);
-    public List<float> rotation = new List<float>(4);
-
-    public void FromGameObject(GameObject gameObject)
-    {
-        position = CoordinateSystem.ToHabitatVector(gameObject.transform.position);
-        rotation = CoordinateSystem.ToHabitatQuaternion(gameObject.transform.rotation);
     }
 }
